moved calculations of yearly geometric means to compiling_data.R
This was computationally heavy, and should only be done once for all analyses so it's the same. This justifies moving it into the compiling data file. However, I saved the results all as a Rdata file to ensure that they are always loaded together. This reduces portability outside of R, but this analysis depends on R anyway.