added script to calculate functional dispersion plot
Also added a new function to the functions script. It's a simple copy-paste of the dbFD function from the FD package, but it gets rid of the annoying requirement to cut trees in interactively instead of via an argument to the function.

This means the functional diversity script can be run via source.