separated conversion factor analysis to create an additional supplemental plot
Prior to this, this branch was loading the conversion factors and applying them to all off the analyses. This reverts back to the original data for the bulk of the analyses. Now the conversion factor script simply makes a new figure illustrating the effect of the conversion factors on a couple representative analyses.